+++ app/scripts/dimApp.config.js
@@ -179,13 +179,8 @@
     .config([
       '$compileProvider',
       function($compileProvider) {
-        var currentImgSrcSanitizationWhitelist = $compileProvider.imgSrcSanitizationWhitelist();
-        var newImgSrcSanitizationWhiteList = currentImgSrcSanitizationWhitelist.toString()
-          .slice(0, -1) + '|chrome-extension:' + currentImgSrcSanitizationWhitelist.toString()
-          .slice(-1);
-
-        //console.log("Changing imgSrcSanitizationWhiteList from " + currentImgSrcSanitizationWhitelist + " to " + newImgSrcSanitizationWhiteList);
-        $compileProvider.imgSrcSanitizationWhitelist(newImgSrcSanitizationWhiteList);
+        // Allow chrome-extension: URLs in ng-src
+        $compileProvider.imgSrcSanitizationWhitelist(/^\s*((https?|chrome-extension):|data:image\/)/);
       }
     ])
     .config(["rateLimiterConfigProvider", function(rateLimiterConfigProvider) {